refactor(card): extract interfaces, standardize file naming and add  prop to CardSection
